pyproject.toml
# SPDX-FileCopyrightText: 2024 JWP Consulting GK
#
# SPDX-License-Identifier: AGPL-3.0-or-later

[project]
name = "projectify"
version = "2025.12.4"
description = "Projectify Backend"
readme = "README.md"
license = "AGPL-3.0-or-later"
license-files = [
    "LICENSES/*"
]
authors = [
    { name = "The Projectify Authors", email = "hello@projectifyapp.com" },
]
maintainers = [
    { name = "Justus Perlwitz", email = "justus@jwpconsulting.net" },
]

[tool.poetry]
name = "projectify"
version = "2025.12.4"
packages = [
    { include = "projectify" },
    { include = "configurations" },
]
exclude = [
    "projectify/*/test",
]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.dependencies]
celery = {extras = ["redis"], version = "~5.4"}
cryptography = "^44.0.1"
dj-database-url = "~2.1.0"
django = "^6.0.0"
django-anymail = {extras = ["mailgun"], version = "~10.3"}
django-celery-results = "~2.5.0"
django-cloudinary-storage = "~0.3.0"
django-pgtrigger = "~4.11.0"
djangorestframework = "^3"
gunicorn = "^23"
pillow = "^11"
psycopg = {version = "^3.1.18", extras = ["c"]}
python = "~3.12.7"
redis = "^5"
rules = "^3.3"
stripe = "^8"
uvicorn = "~0.17.6"
whitenoise = "~6.2.0"
django-ratelimit = "^4.1.0"
django-tailwind = {extras = ["reload"], version = "^3.8.0"}
django-markdownify = "^0.9.5"

[tool.poetry.group.dev]

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1"
faker = "^19.13.0"
django-extensions = "~3.2.0"
# Strictly only used for local development
django-debug-toolbar = "^4.3.0"

ipython = "~8.22"
ruff = "0.6.1"
selenium = "^4.40.0"

[tool.poetry.group.test]

[tool.poetry.group.test.dependencies]
celery-types = "~0.20.0"
django-test-migrations = "^1.3"
django-types = "~0"
mypy = "^1.10"
pyright = "~1.1.318"
pytest = "^8"
pytest-celery = "~0.0.0"
pytest-cov = "^5.0.0"
pytest-django = "~4.5.2"
pytest-xdist = {extras = ["psutil"], version = "~3.3.1"}
types-redis = "~4.6.0.5"
vulture = "^2.9.1"
types-requests = "^2.32.0.20240712"
djlint = "^1.36.1"
types-markdown = "^3.10.2.20260211"
reuse = "^6.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length=79
extend-exclude = [
    # Nix build
    "result/",
    # Vendored in django-configurations
    "configurations",
]

[tool.ruff.lint]
extend-select = [
    "I",
    "E",
    "D",
    "C901",
    # TODO add
    # "C",
    "F",
    # TODO add
    # "DJ",
    # TODO add
    # "T20",
    # There are so many more out there, they all look interesting:
    # https://docs.astral.sh/ruff/rules/
]
ignore = [
    # Conflicting with D201
    "D203",
    # Conflicting with D213
    "D212",
    "E501",
]

[tool.ruff.lint.mccabe]
# Upgrading ruff recently made this metric a lot more sensitive
# I therefore increase it to 12 to make tests pass
# and will at some point decrease complexity in the affected
# again.
max-complexity = 12

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]
split-on-trailing-comma = false
known-first-party = [
    "configurations",
    "projectify",
    "pytest_types",
]

[tool.ruff.lint.isort.sections]
django = [
    "django"
]

[tool.mypy]
mypy_path = [
    "mypy/stubs",
]
strict = true
exclude = [
    ".git",
    "__pycache__",
    "docs",
    "whitelist.py",
    ".direnv",
    "result",
]

# Library imports
[[tool.mypy.overrides]]
module = [
    "cloudinary.*",
    "pgtrigger.*",
]
ignore_missing_imports = true

# Our own modules
[[tool.mypy.overrides]]
module = [
    "projectify.premail.tasks",
]
disallow_untyped_decorators = false


[[tool.mypy.overrides]]
module = [
    "projectify.workspace.test.test_migrations",
    "projectify.premail.test.*",
]
disallow_untyped_defs = false

[tool.pyright]
stubPath = "mypy/stubs"
typeCheckingMode = "strict"
typeshedPath = ""
exclude = [
    ".venv",
    ".direnv",
    "__pycache__/",
    # TODO consider just removing whitelist.py
    "whitelist.py",
    # Created by django so we don't bother
    "projectify/*/migrations/*.py",
    # Missing type stub
    "projectify/workspace/test/test_migrations.py",
    # Type confusion over next(filter())
    "projectify/workspace/views/project.py",
    # This one is sort of covered by mypy
    "projectify/settings/monkeypatch.py",
    # TODO cloudinary stub
    "projectify/utils.py",
    # TODO django rules stub
    "projectify/*/rules.py",
    # TODO write typings
    "projectify/premail/views.py",
    "projectify/premail/test/test_views.py",
    "projectify/premail/test/test_emails.py",
    # TODO false positive in base_url override
    "projectify/storage.py",
    # Nix build
    "result",
    # Pyright reports a few unknown types here that mypy doesn't catch
    "projectify/workspace/views/task.py",
    # Pyright thinks that Deferrable.DEFERRED, a string, is not assignable
    # to Deferrable | None
    "projectify/workspace/models/task.py",
    # CheckConstraint false positive, in mypy we can ignore call-arg
    # pyright emits reportCallIssue instead
    "projectify/workspace/models/workspace.py",
    "projectify/workspace/models/label.py",
    "projectify/user/models/user.py",
    # Other issues
    "projectify/workspace/models/section.py",
    "projectify/workspace/models/sub_task.py",
    # Vendored in
    "configurations/**/*.py",

]
# Mypy doesn't complain about these, so it's fine
reportUnknownParameterType = true
reportPrivateUsage = false
reportIncompatibleVariableOverride = false
reportTypeCommentUsage = false
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnusedImport = false


[tool.vulture]
ignore_decorators = ["@receiver", "@admin.register"]
exclude = [
    ".venv",
    "apps.py",
    "conftest.py",
    "settings",
    "migrations",
    "result/",
    "configurations/",
]
min_confidence = 70

[tool.coverage.run]
include=["projectify/**/*.py"]

[tool.coverage.report]
skip_covered = true
skip_empty = true
show_missing = true
sort = "-cover"

[tool.coverage.html]

[tool.pytest.ini_options]
pythonpath = [
    "projectify",
]
norecursedirs = [
    "result",
    ".direnv",
]
FAIL_INVALID_TEMPLATE_VARS = "True"
DJANGO_SETTINGS_MODULE = "projectify.settings.test"
DJANGO_CONFIGURATION = "Test"
junit_family = "xunit1"
filterwarnings = ["error"]
